Added home controller
