+++ MyMoney.Web/ClientApp/src/app/shared/api/index.ts
@@ -4,3 +4,4 @@ export * from './dtos.interface';
 export * from './income.api';
 export * from './authentication.api';
 export * from './jwt-interceptor.class';
+export * from './home.api';
